Refactor threads filtering + Enable multi-room commenting
